New Update
**WolfBot 6.7.0**

*Finally a NEW UPDATE!*
After a very long hiatus of 3-4 months i update the project.
I made a lot of changes to the Bot since the last time i updated this Project.

# New Things
First of all
I started creating the Developer Commands / Author Commands *<wAuth!>*. This commands are made for make some changes via Discord without need of use VSCode
This commands / functions / schemas are still in ideas, no file has been created yet

# Music System
Changes made to some commands of the Music System.

Lyrics.js
Now is working and shows the lyrics of the song correctly

nowPlaying.js
Finally Added the Progress Bar
Later some comments will be added and the unused code will be removed in the next update

Volume.js
Just commented an unused line

# Slash Commands
Fixing bugs to some Slash Commands *(/src/slashcmd)*.

Added a new Slash Command (Still in Work)

Improved the slashcommands.js File (added more functionality and show errors and help debugging. is not 100% complete
but is very functional)

Remaking the play Slash Command (play.js) because is very buggy and is not functional

Fixing the inventory (inventory.js) is very bugged and don't work after the database was moved

And the Cryptography system was improved (encrypt.js) and now is semi-functional

# Help Prefix Command / Slash Command
The help command and Slash Command are still incomplete, in the next version will be 100% functional

# Index.js
No big changes has been made to this file.

Added a new Embed when a Song is playing
Added more Comments to the code
Added some TODO (To Do) for remember what are the things to improve the next version

# Documents Folder
I deside to add this folder to keep more order in the txt files. More will be added

# .gitkeep
This files are just for create the /dev/ folders

# Packages
Updated some packages and changed others

# Scrpts
Added a fix for no add txt files in the build
Added a Script for delete the Build